pyproject.toml
[project]
name = "receipt-inspector"
version = "0.1.0"
description = "AI receipt validation system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.92.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
